+++ src/StatusMonitor.hpp
@@ -4,20 +4,22 @@
 
 using namespace geode::prelude;
 
-class StatusMonitor : public CCMenu {
+class StatusMonitor : public CCMenu
+{
 protected:
     bool init() override;
     void checkInternetStatus();
     void checkBoomlingsStatus();
     void checkGeodeStatus();
-    
-    CCSprite* m_icon = nullptr;
-    
+
+    CCSprite *m_icon = nullptr;
+
     bool m_internet_ok = false;
     bool m_boomlings_ok = false;
     bool m_geode_ok = false;
+
 public:
     void onEnter() override;
-    static StatusMonitor* create();
+    static StatusMonitor *create();
     void updateStatus(float);
 };