+++ src/StatusPopup.cpp
@@ -7,9 +7,11 @@
 
 using namespace geode::prelude;
 
-StatusPopup* StatusPopup::create() {
+StatusPopup *StatusPopup::create()
+{
     auto ret = new StatusPopup();
-    if (ret && ret->initAnchored(300.f, 180.f)) {
+    if (ret && ret->initAnchored(300.f, 180.f))
+    {
         ret->autorelease();
         return ret;
     }
@@ -17,13 +19,14 @@ StatusPopup* StatusPopup::create() {
     return nullptr;
 }
 
-bool StatusPopup::setup() {
+bool StatusPopup::setup()
+{
     setTitle("Servers Status");
     auto winSize = CCDirector::sharedDirector()->getWinSize();
 
     // user internet
     userInternet = CCLabelBMFont::create("Internet Status: Checking...", "bigFont.fnt");
-    userInternet->setColor({ 100, 100, 100 });
+    userInternet->setColor({100, 100, 100});
     userInternet->setPosition(m_mainLayer->getContentSize().width / 2, m_mainLayer->getContentSize().height / 2 + 30);
     userInternet->setScale(0.5f);
     m_mainLayer->addChild(userInternet);
@@ -40,7 +43,7 @@ bool StatusPopup::setup() {
     // display boomlings server status
     serverStatus = CCLabelBMFont::create("Boomlings Status: Checking...", "bigFont.fnt");
     serverStatus->setPosition(m_mainLayer->getContentSize().width / 2, m_mainLayer->getContentSize().height / 2);
-    serverStatus->setColor({ 100, 100, 100 });
+    serverStatus->setColor({100, 100, 100});
     serverStatus->setScale(0.5f);
     m_mainLayer->addChild(serverStatus);
     // subtext timestamp
@@ -55,7 +58,7 @@ bool StatusPopup::setup() {
 
     // display geode server status
     geodeStatus = CCLabelBMFont::create("GeodeSDK Status: Checking...", "bigFont.fnt");
-    geodeStatus->setColor({ 100, 100, 100 });
+    geodeStatus->setColor({100, 100, 100});
     geodeStatus->setPosition(m_mainLayer->getContentSize().width / 2, m_mainLayer->getContentSize().height / 2 - 30);
     geodeStatus->setScale(0.5f);
     m_mainLayer->addChild(geodeStatus);
@@ -72,20 +75,18 @@ bool StatusPopup::setup() {
     // mod settings button
     auto modSettingsMenu = CCMenu::create();
     modSettingsMenu->setPosition({0, 0});
-        auto modSettingsBtnSprite = CircleButtonSprite::createWithSpriteFrameName(
+    auto modSettingsBtnSprite = CircleButtonSprite::createWithSpriteFrameName(
         // @geode-ignore(unknown-resource)
         "geode.loader/settings.png",
         1.f,
         CircleBaseColor::Green,
-        CircleBaseSize::Medium
-    );
+        CircleBaseSize::Medium);
     modSettingsBtnSprite->setScale(0.75f);
 
     auto modSettingsButton = CCMenuItemSpriteExtra::create(
         modSettingsBtnSprite,
         this,
-        menu_selector(StatusPopup::onModSettings)
-    );
+        menu_selector(StatusPopup::onModSettings));
     modSettingsMenu->addChild(modSettingsButton);
     m_mainLayer->addChild(modSettingsMenu);
 
@@ -94,9 +95,12 @@ bool StatusPopup::setup() {
     checkBoomlingsStatus();
     checkGeodeStatus();
     // immediate status refresh on the existing monitor instance
-    if (auto scene = CCDirector::sharedDirector()->getRunningScene()) {
-        if (auto node = scene->getChildByID("status-monitor")) {
-            if (auto monitor = typeinfo_cast<StatusMonitor*>(node)) {
+    if (auto scene = CCDirector::sharedDirector()->getRunningScene())
+    {
+        if (auto node = scene->getChildByID("status-monitor"))
+        {
+            if (auto monitor = typeinfo_cast<StatusMonitor *>(node))
+            {
                 monitor->updateStatus(0.0f);
             }
         }
@@ -105,26 +109,31 @@ bool StatusPopup::setup() {
     return true;
 }
 
-void StatusPopup::onModSettings(CCObject* sender) {
+void StatusPopup::onModSettings(CCObject *sender)
+{
     openSettingsPopup(getMod());
 }
 
-void StatusPopup::checkInternetStatus() {
+void StatusPopup::checkInternetStatus()
+{
     log::debug("checking internet status");
     // do we have internet?
-    if (Mod::get()->getSettingValue<bool>("doWeHaveInternet")) {
-        if (GameToolbox::doWeHaveInternet()) {
+    if (Mod::get()->getSettingValue<bool>("doWeHaveInternet"))
+    {
+        if (GameToolbox::doWeHaveInternet())
+        {
             userInternet->setString("Internet Status: Online");
-            userInternet->setColor({ 0, 255, 0 });
+            userInternet->setColor({0, 255, 0});
             return;
         }
         log::debug("Internet offline or unreachable");
         userInternet->setString("Internet Status: Offline");
-        userInternet->setColor({ 255, 0, 0 });
+        userInternet->setColor({255, 0, 0});
         return;
     }
     auto url = Mod::get()->getSettingValue<std::string>("internet_url");
-    geode::utils::web::WebRequest().get(url).listen([this, url](geode::utils::web::WebResponse* response) {
+    geode::utils::web::WebRequest().get(url).listen([this, url](geode::utils::web::WebResponse *response)
+                                                    {
         if (!response || !response->ok()) {
             log::debug("{} offline or unreachable", url);
             if (userInternet) {
@@ -137,32 +146,40 @@ void StatusPopup::checkInternetStatus() {
         if (userInternet) {
             userInternet->setString("Internet Status: Online");
             userInternet->setColor({ 0, 255, 0 });
-        }
-    });
+        } });
 }
 
-void StatusPopup::checkBoomlingsStatus() {
+void StatusPopup::checkBoomlingsStatus()
+{
     log::debug("checking Boomlings server status");
-    geode::utils::web::WebRequest().get("https://www.boomlings.com").listen([this](geode::utils::web::WebResponse* response) {
-        if (!response || !response->ok()) {
-            log::debug("Boomlings server offline or unreachable");
-            if (serverStatus) {
-                serverStatus->setString("Boomlings Status: Offline");
-                serverStatus->setColor({ 255, 0, 0 });
+    auto url = "http://www.boomlings.com/database/getGJLevels21.php";
+    auto body = "type=2&secret=Wmfd2893gb7"; // most liked level
+
+    geode::utils::web::WebRequest()
+        .bodyString(body)
+        .post(url)
+        .listen([this](geode::utils::web::WebResponse *response)
+                {
+            if (!response || !response->ok()) {
+                log::debug("Boomlings server offline or unreachable");
+                if (serverStatus) {
+                    serverStatus->setString("Boomlings Status: Offline");
+                    serverStatus->setColor({255, 0, 0});
+                }
+                return;
             }
-            return;
-        }
-        log::debug("Boomlings server online");
-        if (serverStatus) {
-            serverStatus->setString("Boomlings Status: Online");
-            serverStatus->setColor({ 0, 255, 0 });
-        }
-    });
+            log::debug("Boomlings server online");
+            if (serverStatus) {
+                serverStatus->setString("Boomlings Status: Online");
+                serverStatus->setColor({0, 255, 0});
+            } });
 }
 
-void StatusPopup::checkGeodeStatus() {
+void StatusPopup::checkGeodeStatus()
+{
     log::debug("checking GeodeSDK status");
-    geode::utils::web::WebRequest().get("https://api.geode-sdk.org").listen([this](geode::utils::web::WebResponse* response) {
+    geode::utils::web::WebRequest().get("https://api.geode-sdk.org").listen([this](geode::utils::web::WebResponse *response)
+                                                                            {
         if (!response || !response->ok()) {
             log::debug("GeodeSDK offline or unreachable");
             if (geodeStatus) {
@@ -175,6 +192,5 @@ void StatusPopup::checkGeodeStatus() {
         if (geodeStatus) {
             geodeStatus->setString("GeodeSDK Status: Online");
             geodeStatus->setColor({ 0, 255, 0 });
-        }
-    });
+        } });
 };